Add files via upload
In this notebook, we explore the data about independent researchers as gathered from Open Alex. The notebook gives an overview of some of the main characteristics of publications authored by independent researchers: the topics, the countries, the number of co-authors... There is also a section on the open access characteristics of publications by independent researchers.